refactor: 💡 code imp
